housing: fe refactor + start integration with be
