+++ frontend/package.json
@@ -15,6 +15,7 @@
 	},
 	"devDependencies": {
 		"@types/d3": "^7.4.3",
+		"@zerodevx/svelte-toast": "^0.9.6",
 		"prettier": "^3.4.2",
 		"prettier-plugin-svelte": "^3.3.3",
 		"svelte": "^5.34.3",
