+++ frontend/src/lib/housing/components/layout/Footer.svelte
