+++ frontend/src/lib/housing/components/submit-review/Footer.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+</script>
+
+<div class="mt-8 text-center text-sm text-gray-500 dark:text-gray-400">
+	<p>Reviews are anonymous and help create a better housing experience for all Purdue students.</p>
+</div>
