+++ frontend/src/lib/housing/components/submit-review/Header.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+</script>
+
+<div class="text-center mb-8">
+	<h2 class="text-3xl font-bold text-gray-900 dark:text-gray-100 mb-4">
+		Share Your Housing Experience
+	</h2>
+	<p class="text-lg text-gray-600 dark:text-gray-400">
+		Help fellow Boilermakers make informed housing decisions with your honest review
+	</p>
+</div>
