+++ frontend/src/lib/housing/components/submit-review/subcomponents/Description.svelte
@@ -1,0 +1,27 @@
+<script lang="ts">
+	import { PUBLIC_HOUSING_MAX_CHARS, PUBLIC_HOUSING_MIN_CHARS } from '$env/static/public'
+
+	let comment = ''
+</script>
+
+<div>
+	<div
+		class="leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70 text-base font-semibold dark:text-gray-200"
+	>
+		Your Review *
+		<p class="text-sm text-gray-600 dark:text-gray-400 mb-3">Share details about your experience</p>
+		<textarea
+			id="comment"
+			bind:value={comment}
+			placeholder="Tell us about your experience living here. What did you like? What could be improved? Any tips for future residents?"
+			rows="6"
+			required
+			maxlength={Number(PUBLIC_HOUSING_MAX_CHARS)}
+			minlength={Number(PUBLIC_HOUSING_MIN_CHARS)}
+			class="flex min-h-[80px] w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 resize-none dark:bg-gray-700 dark:border-gray-600 dark:text-gray-200"
+		></textarea>
+		<p class="text-xs text-gray-500 dark:text-gray-400 mt-1">
+			Minimum {PUBLIC_HOUSING_MIN_CHARS} characters ({comment.length}/{PUBLIC_HOUSING_MIN_CHARS})
+		</p>
+	</div>
+</div>
